Apply new generation gwt-maven-plugin
