--- .factorypath
@@ -1,9 +1,0 @@
-<factorypath>
-    <factorypathentry kind="PLUGIN" id="com.ibm.jee.annotations.processor" enabled="true" runInBatchMode="false"/>
-    <factorypathentry kind="PLUGIN" id="com.ibm.jaxrs.annotations.processor" enabled="true" runInBatchMode="false"/>
-    <factorypathentry kind="PLUGIN" id="com.ibm.etools.javaee.cdi.ext.ui" enabled="true" runInBatchMode="false"/>
-    <factorypathentry kind="VARJAR" id="M2_REPO\com\google\dagger\dagger\2.6\dagger-2.6.jar" enabled="true" runInBatchMode="false"/>
-    <factorypathentry kind="VARJAR" id="M2_REPO\com\google\dagger\dagger-compiler\2.6\dagger-compiler-2.6.jar" enabled="true" runInBatchMode="false"/>
-    <factorypathentry kind="VARJAR" id="M2_REPO\com\google\dagger\dagger-producers\2.6\dagger-producers-2.6.jar" enabled="true" runInBatchMode="false"/>    
-    <factorypathentry kind="VARJAR" id="M2_REPO\com\google\guava\guava\19.0\guava-19.0.jar" enabled="true" runInBatchMode="false"/>
-</factorypath>
