--- res/VERSION
@@ -1,1 +1,0 @@
-${project.version}
