--- src/com/example/gwt/dagger2/client/di/CoffeeApp.java
@@ -1,21 +1,0 @@
-package com.example.gwt.dagger2.client.di;
-
-import dagger.Component;
-import javax.inject.Singleton;
-
-public class CoffeeApp {
-
-	@Singleton
-	@Component(modules = { DripCoffeeModule.class })
-	public interface Coffee {
-		
-		CoffeeMaker maker();
-	
-	}
-//
-//	public static void main(String[] args) {
-//		Coffee coffee = DaggerCoffeeApp_Coffee.builder().build();
-//		coffee.maker().brew();
-//	}
-
-}
