--- src/com/example/gwt/dagger2/client/di/Heater.java
@@ -1,11 +1,0 @@
-package com.example.gwt.dagger2.client.di;
-
-interface Heater {
-	
-	void on();
-
-	void off();
-
-	boolean isHot();
-
-}
