--- src/com/example/gwt/dagger2/client/di/Pump.java
@@ -1,7 +1,0 @@
-package com.example.gwt.dagger2.client.di;
-
-interface Pump {
-	
-	void pump();
-
-}
