+++ src/main/java/coffee/client/DripCoffeeModule.java
@@ -1,4 +1,4 @@
-package com.example.gwt.dagger2.client.di;
+package coffee.client;
 
 import dagger.Module;
 import dagger.Provides;
@@ -6,11 +6,7 @@ import javax.inject.Singleton;
 
 @Module(includes = PumpModule.class)
 class DripCoffeeModule {
-
-	@Provides
-	@Singleton
-	Heater provideHeater() {
-		return new ElectricHeater();
-	}
-
+    @Provides @Singleton Heater provideHeater() {
+        return new ElectricHeater();
+    }
 }
