+++ src/main/java/coffee/client/Heater.java
@@ -1,0 +1,7 @@
+package coffee.client;
+
+interface Heater {
+    void on();
+    void off();
+    boolean isHot();
+}
