+++ src/main/java/coffee/client/Pump.java
@@ -1,0 +1,5 @@
+package coffee.client;
+
+interface Pump {
+    void pump();
+}
