+++ src/main/java/coffee/client/PumpModule.java
@@ -1,0 +1,10 @@
+package coffee.client;
+
+import dagger.Binds;
+import dagger.Module;
+
+@Module
+abstract class PumpModule {
+    @Binds
+    abstract Pump providePump(Thermosiphon pump);
+}
