+++ src/main/resources/coffee/CoffeeApp.gwt.xml
@@ -1,11 +1,5 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  When updating your version of GWT, you should also update this DTD reference,
-  so that your app can take advantage of the latest GWT module capabilities.
--->
-<!DOCTYPE module PUBLIC "-//Google Inc.//DTD Google Web Toolkit 2.8.0//EN"
-  "http://gwtproject.org/doctype/2.8.0/gwt-module.dtd">
-<module rename-to="dagger2">
+<!DOCTYPE module PUBLIC "-//Google Inc.//DTD Google Web Toolkit 2.7.0//EN" "http://gwtproject.org/doctype/2.7.0/gwt-module.dtd">
+<module rename-to="coffee">
   <!-- Inherit the core Web Toolkit stuff.                        -->
   <inherits name="com.google.gwt.user.User"/>
 
@@ -21,17 +15,11 @@
   <inherits name="dagger.Dagger" />
   
   <!-- Specify the app entry point class.                         -->
-  <entry-point class="com.example.gwt.dagger2.client.Dagger2"/>
+  <entry-point class="coffee.client.CoffeeApp"/>
+
+  <define-configuration-property name="project.version" is-multi-valued="false"/>
 
   <!-- Specify the paths for translatable code                    -->
   <source path="client"/>
-
-  <!-- allow Super Dev Mode -->
-  <add-linker name="xsiframe"/>
-
-  <!-- Generators -->
-  <generate-with class="com.example.gwt.dagger2.rebind.VersionGenerator">
-    <when-type-assignable class="com.example.gwt.dagger2.client.Version"/>
-  </generate-with>  
-  
+  <public path="public"/>
 </module>
