+++ src/main/resources/coffee/public/favicon.png
