+++ src/main/resources/coffee/public/index.html
@@ -4,9 +4,8 @@
     <meta http-equiv="content-type" content="text/html; charset=UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="icon" href="favicon.png" type="image/png">
-    <link type="text/css" rel="stylesheet" href="default.css">
-    <title>GWT+Dagger2 Example</title>
-    <script type="text/javascript" src="dagger2/dagger2.nocache.js"></script>
+    <title>GWT+Dagger2 Coffee Example</title>
+    <script type="text/javascript" src="coffee.nocache.js"></script>
   </head>
   <body>
     <!-- RECOMMENDED if your web app will not function without JavaScript enabled -->
