--- war/default.css
@@ -1,1 +1,0 @@
-/** Add css rules here for your application. */
